+++ src/main/java/cloudcomputinginha/demo/web/controller/ResumeRestController.java
@@ -6,9 +6,9 @@ import cloudcomputinginha.demo.converter.ResumeConverter;
 import cloudcomputinginha.demo.domain.Interview;
 import cloudcomputinginha.demo.domain.Resume;
 import cloudcomputinginha.demo.infra.ai.AiOpenFeignService;
+import cloudcomputinginha.demo.infra.aws.ResumeS3Service;
 import cloudcomputinginha.demo.service.resume.ResumeCommandService;
 import cloudcomputinginha.demo.service.resume.ResumeQueryService;
-import cloudcomputinginha.demo.service.resume.ResumeS3Service;
 import cloudcomputinginha.demo.validation.annotation.ValidFileName;
 import cloudcomputinginha.demo.web.dto.InterviewResponseDTO;
 import cloudcomputinginha.demo.web.dto.ResumeRequestDTO;
